% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixoncoles.R
\name{augment.dixoncoles}
\alias{augment.dixoncoles}
\title{Augment data with information from a Dixon-Coles model}
\usage{
augment.dixoncoles(x, data = NULL, newdata, type.predict, ...)
}
\arguments{
\item{x}{A `dixoncoles` object created by `regista::dixoncoles`.}

\item{data}{A `data.frame` or `tibble::tibble` containing the original data.}

\item{newdata}{A `data.frame` or `tibble::tibble` object of new data to be predicted.}

\item{type.predict}{Type of prediction. Passed onto `regista::predict.dixoncoles`}

\item{...}{Additional arguments. Not used.}
}
\value{
A `tibble::tibble()` with one row.
}
\description{
Append additional information about a set of matches.
}
