% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixoncoles.R
\name{dixoncoles}
\alias{dixoncoles}
\title{Dixon-Coles model for estimating team strengths}
\usage{
dixoncoles(hgoal, agoal, hteam, ateam, data, weights = ~1, ...)
}
\arguments{
\item{hgoal}{A formula describing the home goals column in `data`, or a
numeric vector containing the observed home goals for a set of games.}

\item{agoal}{A formula describing the away goals column in `data`, or a
numeric vector containing the observed away goals for a set of games.}

\item{hteam}{A formula describing the home team column in `data`, or a
vector containing the home team name for a set of games.}

\item{ateam}{A formula describing the away team column in `data`, or a
vector containing the away team name for a set of games.}

\item{data}{Data frame, list or environment (or object coercible by
`as.data.frame` to a data frame) containing the variables in the model.}

\item{weights}{A formula describing an expression to calculate the weight for
each game. All games weighted equally by default.}

\item{...}{Arguments passed onto `dixoncoles_ext`.}
}
\value{
A list with component `par` containing the best set of parameters
  found. See `optim` for details.
}
\description{
This is an implementation of the Dixon-Coles model for estimating soccer
teams' strength from goals scored and conceded:

Dixon, Mark J., and Stuart G. Coles. "Modelling association football scores
and inefficiencies in the football betting market." Journal of the Royal
Statistical Society: Series C (Applied Statistics) 46, no. 2 (1997):
265-280.
}
\examples{
fit <- dixoncoles(~hgoal, ~agoal, ~home, ~away,
                  data = premier_league_2010)

}
