% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixoncoles.R
\name{.dc_modeldata}
\alias{.dc_modeldata}
\title{Get model data for a Dixon-Coles model}
\usage{
.dc_modeldata(f1, f2, weights, data, predict = FALSE)
}
\description{
Get model data for a Dixon-Coles model
}
\keyword{internal}
