% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixoncoles.R
\name{.dc_negloglike}
\alias{.dc_negloglike}
\title{Dixon-Coles negative log likelihood}
\usage{
.dc_negloglike(hg, ag, home_rates, away_rates, rho, weights)
}
\description{
Dixon-Coles negative log likelihood
}
\keyword{internal}
