% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixoncoles.R
\name{.dc_objective_function}
\alias{.dc_objective_function}
\title{Dixon-Coles objective function}
\usage{
.dc_objective_function(params, modeldata)
}
\description{
Dixon-Coles objective function
}
\keyword{internal}
