% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixoncoles.R
\name{.dc_predict_scorelines}
\alias{.dc_predict_scorelines}
\title{Calculate the probability of scorelines occuring for a given set of matches}
\usage{
.dc_predict_scorelines(rates, up_to, threshold)
}
\description{
Calculate the probability of scorelines occuring for a given set of matches
}
\keyword{internal}
