% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixoncoles.R
\name{.dc_predict_scorelines_once}
\alias{.dc_predict_scorelines_once}
\title{Calculate the probability of scorelines occuring for a given match}
\usage{
.dc_predict_scorelines_once(home_rate, away_rate, rho, up_to, threshold)
}
\description{
Calculate the probability of scorelines occuring for a given match
}
\keyword{internal}
