% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixoncoles.R
\name{.dc_rate_info}
\alias{.dc_rate_info}
\title{Get estimated rates for home and away goals}
\usage{
.dc_rate_info(params, modeldata)
}
\description{
Get estimated rates for home and away goals
}
\keyword{internal}
