% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixoncoles.R
\name{.fill_if_missing}
\alias{.fill_if_missing}
\title{Add column to a matrix, if it doesn't exist}
\usage{
.fill_if_missing(mat, name)
}
\description{
Add column to a matrix, if it doesn't exist
}
\keyword{internal}
