% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixoncoles.R
\name{.make_dummies}
\alias{.make_dummies}
\title{Get a matrix of dummy variables from a factor}
\usage{
.make_dummies(values)
}
\description{
Get a matrix of dummy variables from a factor
}
\keyword{internal}
