% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixoncoles.R
\name{.normalise_off_params}
\alias{.normalise_off_params}
\title{Normalise attack parameters to make the model identifable (mean = 1)}
\usage{
.normalise_off_params(params)
}
\description{
Normalise attack parameters to make the model identifable (mean = 1)
}
\keyword{internal}
