% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixoncoles.R
\name{.tau}
\alias{.tau}
\title{Function controlling dependence between home and away goals}
\usage{
.tau(hg, ag, home_rates, away_rates, rho)
}
\description{
Function controlling dependence between home and away goals
}
\keyword{internal}
