% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_teams.R
\name{factor_teams}
\alias{factor_teams}
\title{Coerce teamnames into factors}
\usage{
factor_teams(data, columns)
}
\arguments{
\item{data}{A data frame containing match data.}

\item{columns}{A charater vector containing the names of columns
containing team names.}
}
\description{
Modelling team strengths often requires separate parameters for offence and
defence. These helper functions provide a quick way to turn a factor of team
names into a factor of offence or defence levels
}
\examples{
# For a given dataframe of matches...
games <- data.frame(
  home = c("Arsenal", "Barcelona", "Arsenal"),
  away = c("Chelsea", "Roma",      "Barcelona"),
  result = c("H", "A", "H"),
  stringsAsFactors = FALSE
)

# Convert the teams to factors
games <- factor_teams(games, c("home", "away"))
levels(games$home)
#> c("Arsenal", "Barcelona", "Chelsea", "Roma")
}
