% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{off}
\alias{off}
\alias{def}
\title{Helper functions to enable easier model specification}
\usage{
off(x)

def(x)
}
\arguments{
\item{x}{A factor containing team names}
}
\description{
Modelling team strengths often requires separate parameters for offence and
defence. These helper functions provide a quick way to turn a factor of team
names into a factor of offence or defence levels
}
\examples{
teams <- factor(c("Arsenal", "Liverpool", "Barcelona", "Dortmund"))
team_off <- off(teams)

}
