% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixoncoles.R
\name{predict.dixoncoles}
\alias{predict.dixoncoles}
\title{Predict method for Dixon-Coles model fits}
\usage{
\method{predict}{dixoncoles}(object, newdata, type = c("rates", "scorelines"),
  up_to = 50, threshold = 1e-08, ...)
}
\arguments{
\item{object}{Object of class inheriting from `dixoncoles`.}

\item{newdata}{A data frame in which to look for variables to predict}

\item{type}{Type of prediction (rates or scorelines).}

\item{up_to}{If `type = "scorelines"`, the maximum number of goals for which
to calculate the probability of occurring in each match.}

\item{threshold}{If `type = "scorelines"`, scorelines with a probability
below `threshold` will not be returned.}

\item{...}{Arguments passed from other methods}
}
\value{
If `type = "rates"`, a dataframe of each teams' estimated goalscoring
rate. If `type = "scorelines"`, a list in which each element is a dataframe
of scorelines and their estimated probabilities.
}
\description{
Predicted rates or scorelines based on a Dixon Coles model object
}
