% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{premier_league_2010}
\alias{premier_league_2010}
\title{Results for the 2010/11 English Premier League season}
\format{A data frame with 380 rows and 6 variables:
\describe{
  \item{date}{date that the game was played}
  \item{home}{The home team}
  \item{away}{The away team}
  \item{hgoal}{The number of goals scored by the home team}
  \item{agoal}{The number of goals scored by the away team}
  \item{result}{The match result (H/D/A)}
  \item{hfa}{Home field advantage (whether the `home` team played at their
             home stadium)}
}}
\usage{
premier_league_2010
}
\description{
A tibble containing the home and away goals for each game in the 2010/11
English Premier League season.
}
\keyword{datasets}
