% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scorelines_to_outcomes}
\alias{scorelines_to_outcomes}
\title{Aggregate scoreline probabilities to 1X2 probabilities}
\usage{
scorelines_to_outcomes(scorelines, hgoal = ~hgoal, agoal = ~agoal,
  prob = ~prob)
}
\description{
Often we are more interested in the outcome of a match (home win, draw,
away win) than we are in the scorelines. This is a helper function that
aggregates scoreline probabilities to outcome probabilities.
}
