% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixoncoles.R
\name{tidy.dixoncoles}
\alias{tidy.dixoncoles}
\title{Tidy a Dixon-Coles model}
\usage{
tidy.dixoncoles(x)
}
\arguments{
\item{x}{A `dixoncoles` object created by `regista::dixoncoles()`}

\item{...}{Additional arguments. Not used.}
}
\value{
A `tibble::tibble()` with one row for each parameter estimated by the
model.
}
\description{
Tidy summarises information about a fitted Dixon-Coles model.
}
